# BlazorAppOrquesta/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SistemaGestionOrquesta",
    "version": "1.0"
  },
  "paths": {
...
    "/estudiante/save": {
      "post": {
        "tags": [
          "Estudiante"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudiante/delete/{id}": {
      "delete": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudiante/baja/{id}": {
      "delete": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudiante/update/{id}": {
      "put": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EstudianteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/estudiante/{id}": {
      "get": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudianteNombreYDni/{nombre}/{documento}": {
      "get": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudianteNombreApellido/{nombre}/{apellido}": {
      "get": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apellido",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudiantesByCurso/{idCurso}": {
      "get": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "idCurso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estudiante"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estudiante"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Estudiante"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cursosByEstudiante/{idEstudiante}": {
      "get": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "idEstudiante",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curso"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curso"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Curso"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/estudianteEliminarCurso/{id}/{Idcurso}": {
      "delete": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idcurso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudianteDarDeAltaEnCurso/{id}/{Idcurso}": {
      "post": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idcurso",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/estudianteCambiarCurso/{id}/{cursoNuevo}/{cursoViejo}": {
      "put": {
        "tags": [
          "Estudiante"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursoNuevo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursoViejo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Curso": {
        "type": "object",
        "properties": {
          "cursoId": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "profesorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Estudiante": {
        "type": "object",
        "properties": {
          "estudianteId": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "documento": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "instrumentoId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rutaFoto": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean",
            "nullable": true
          },
          "nombreTutor": {
            "type": "string",
            "nullable": true
          },
          "telefonoTutor": {
            "type": "string",
            "nullable": true
          },
          "asegurado": {
            "type": "boolean",
            "nullable": true
          },
          "documentoTutor": {
            "type": "string",
            "nullable": true
          },
          "documentoTutor2": {
            "type": "string",
            "nullable": true
          },
          "nombreTutor2": {
            "type": "string",
            "nullable": true
          },
          "telefonoTutor2": {
            "type": "string",
            "nullable": true
          },
          "nacionalidad": {
            "type": "string",
            "nullable": true
          },
          "tmtMédico": {
            "type": "string",
            "nullable": true
          },
          "epPsicoMotriz": {
            "type": "string",
            "nullable": true
          },
          "particularidad": {
            "type": "string",
            "nullable": true
          },
          "autoretiro": {
            "type": "boolean",
            "nullable": true
          },
          "instrumento": {
            "$ref": "#/components/schemas/Instrumento"
          },
          "prestamosInstrumentos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrestamosInstrumento"
            },
            "nullable": true
          },
          "cursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Curso"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EstudianteDTO": {
        "type": "object",
        "properties": {
          "estudianteId": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "documento": {
            "type": "string",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "nacionalidad": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "nombreTutor": {
            "type": "string",
            "nullable": true
          },
          "documentoTutor": {
            "type": "string",
            "nullable": true
          },
          "nombreTutor2": {
            "type": "string",
            "nullable": true
          },
          "documentoTutor2": {
            "type": "string",
            "nullable": true
          },
          "telefonoTutor": {
            "type": "string",
            "nullable": true
          },
          "telefonoTutor2": {
            "type": "string",
            "nullable": true
          },
          "orquesta": {
            "type": "string",
            "nullable": true
          },
          "instrumentoId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "profeCursoViolin": {
            "type": "string",
            "nullable": true
          },
          "rutaFoto": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean",
            "nullable": true
          },
          "asegurado": {
            "type": "boolean",
            "nullable": true
          },
          "tmtMédico": {
            "type": "string",
            "nullable": true
          },
          "epPsicoMotriz": {
            "type": "string",
            "nullable": true
          },
          "particularidad": {
            "type": "string",
            "nullable": true
          },
          "autoretiro": {
            "type": "boolean",
            "nullable": true
          },
          "instrumentoNombre": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Instrumento": {
        "type": "object",
        "properties": {
          "instrumentoId": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "detalles": {
            "type": "string",
            "nullable": true
          },
          "disponible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PrestamosInstrumento": {
        "type": "object",
        "properties": {
          "prestamoInstrumentoId": {
            "type": "integer",
            "format": "int32"
          },
          "fechaPrestamo": {
            "type": "string",
            "format": "date-time"
          },
          "fechaDevolucion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "instrumentoId": {
            "type": "integer",
            "format": "int32"
          },
          "estudianteId": {
            "type": "string",
            "format": "uuid"
          },
          "estudiante": {
            "$ref": "#/components/schemas/Estudiante"
          },
          "instrumento": {
            "$ref": "#/components/schemas/Instrumento"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}